feat: add create sensor function

+++ rust/sistema_florestal/src/constants.rs
@@ -1,3 +1,3 @@
-pub const DISTANCE_WARNING_MESSAGE: i8 = 3;
-pub const DISTANCE_VIEW: i8 = 1;
+pub const DISTANCE_WARNING_MESSAGE: usize = 3;
+pub const DISTANCE_VIEW: usize = 1;
 pub const TAM_MAP_FLOREST: (usize, usize) = (6, 11);
